FloatingAddition - bugfixes, cleanup
The original method of comparing absolute values of numbers is flawed as it
ignores mantissa values. This is relevant in case when exponents are the same.
Comparison logic was extracted to the FloatingComparison module. In order to compare
absolute values, sign is replaced by 0 on the input of this module.

The logic of swapping numbers based on comparison result was also cleared up.
Only values in a_swap, b_swap registers are swapped, the specific parts of a float
are now wires assigned to parts of this register. This is also the reason why
B_shifted_mantissa was created, as wire cannot be reassigned.

Overflow protection to exponent was added in the case of a carry.

The mantissa left-shifting was behaving as an infinite loop in some cases,
so changes had to be made by replacing while loop by a for loop with limited
iterations.

Some excessive unused variables were removed.